Move implicit tendency to src
