+++ examples/hybrid/staggered_nonhydrostatic_model.jl
@@ -124,7 +124,17 @@ function default_cache(Y, params, model_spec, spaces, numerics, simulation)
         :ρe_tot in pnc || :ρe_int in pnc ? (; ᶜρh = similar(Y.c, FT)) : ()
     return (;
         simulation,
-        operators = (; ᶜdivᵥ, ᶠinterp, ᶠcurlᵥ, ᶜinterp, ᶠgradᵥ),
+        operators = (;
+            ᶜdivᵥ,
+            ᶠinterp,
+            ᶠcurlᵥ,
+            ᶜinterp,
+            ᶠgradᵥ,
+            ᶠupwind1,
+            ᶠupwind3,
+            ᶠfct_boris_book,
+            ᶠfct_zalesak,
+        ),
         spaces,
         Yₜ = similar(Y), # only needed when using increment formulation
         limiters,
@@ -156,140 +166,12 @@ function default_cache(Y, params, model_spec, spaces, numerics, simulation)
     )
 end
 
-# TODO: All of these should use dtγ instead of dt, but dtγ is not available in
-# the implicit tendency function. Since dt >= dtγ, we can safely use dt for now.
-vertical_transport!(ᶜρcₜ, ᶠw, ᶜρ, ᶜρc, dt, ::Val{:none}) =
-    @. ᶜρcₜ = -(ᶜdivᵥ(ᶠinterp(ᶜρc) * ᶠw))
-vertical_transport!(ᶜρcₜ, ᶠw, ᶜρ, ᶜρc, dt, ::Val{:first_order}) =
-    @. ᶜρcₜ = -(ᶜdivᵥ(ᶠinterp(ᶜρ) * ᶠupwind1(ᶠw, ᶜρc / ᶜρ)))
-vertical_transport!(ᶜρcₜ, ᶠw, ᶜρ, ᶜρc, dt, ::Val{:third_order}) =
-    @. ᶜρcₜ = -(ᶜdivᵥ(ᶠinterp(ᶜρ) * ᶠupwind3(ᶠw, ᶜρc / ᶜρ)))
-vertical_transport!(ᶜρcₜ, ᶠw, ᶜρ, ᶜρc, dt, ::Val{:boris_book}) = @. ᶜρcₜ =
-    -(ᶜdivᵥ(ᶠinterp(ᶜρ) * ᶠupwind1(ᶠw, ᶜρc / ᶜρ))) - ᶜdivᵥ(
-        ᶠinterp(ᶜρ) * ᶠfct_boris_book(
-            ᶠupwind3(ᶠw, ᶜρc / ᶜρ) - ᶠupwind1(ᶠw, ᶜρc / ᶜρ),
-            (ᶜρc / dt - ᶜdivᵥ(ᶠinterp(ᶜρ) * ᶠupwind1(ᶠw, ᶜρc / ᶜρ))) / ᶜρ,
-        ),
-    )
-vertical_transport!(ᶜρcₜ, ᶠw, ᶜρ, ᶜρc, dt, ::Val{:zalesak}) = @. ᶜρcₜ =
-    -(ᶜdivᵥ(ᶠinterp(ᶜρ) * ᶠupwind1(ᶠw, ᶜρc / ᶜρ))) - ᶜdivᵥ(
-        ᶠinterp(ᶜρ) * ᶠfct_zalesak(
-            ᶠupwind3(ᶠw, ᶜρc / ᶜρ) - ᶠupwind1(ᶠw, ᶜρc / ᶜρ),
-            ᶜρc / ᶜρ / dt,
-            (ᶜρc / dt - ᶜdivᵥ(ᶠinterp(ᶜρ) * ᶠupwind1(ᶠw, ᶜρc / ᶜρ))) / ᶜρ,
-        ),
-    )
-
-# Used for automatically computing the Jacobian ∂Yₜ/∂Y. Currently requires
-# allocation because the cache is stored separately from Y, which means that
-# similar(Y, <:Dual) doesn't allocate an appropriate cache for computing Yₜ.
-function implicit_cache_vars(
-    Y::Fields.FieldVector{T},
-    p,
-) where {T <: AbstractFloat}
-    (; ᶜK, ᶜts, ᶜp) = p
-    return (; ᶜK, ᶜts, ᶜp)
-end
-function implicit_cache_vars(Y::Fields.FieldVector{T}, p) where {T <: Dual}
-    ᶜρ = Y.c.ρ
-    ᶜK = similar(ᶜρ)
-    ᶜts = similar(ᶜρ, eltype(p.ts).name.wrapper{eltype(ᶜρ)})
-    ᶜp = similar(ᶜρ)
-    return (; ᶜK, ᶜts, ᶜp)
-end
-
 function implicit_tendency!(Yₜ, Y, p, t)
     @nvtx "implicit tendency" color = colorant"yellow" begin
         Fields.bycolumn(axes(Y.c)) do colidx
-            implicit_tendency!(Yₜ, Y, p, t, colidx)
+            CA.implicit_tendency!(Yₜ, Y, p, t, colidx)
         end
     end
-end
-
-function implicit_tendency!(Yₜ, Y, p, t, colidx)
-    ᶜρ = Y.c.ρ
-    ᶜuₕ = Y.c.uₕ
-    ᶠw = Y.f.w
-    (; ᶜK, ᶠgradᵥ_ᶜΦ, ᶜts, ᶜp, params, thermo_dispatcher) = p
-    (; energy_upwinding, tracer_upwinding, simulation) = p
-
-    thermo_params = CAP.thermodynamics_params(params)
-    dt = simulation.dt
-    @. ᶜK[colidx] = norm_sqr(C123(ᶜuₕ[colidx]) + C123(ᶜinterp(ᶠw[colidx]))) / 2
-    CA.thermo_state!(Y, p, ᶜinterp, colidx)
-    @. ᶜp[colidx] = TD.air_pressure(thermo_params, ᶜts[colidx])
-
-    if p.tendency_knobs.has_turbconv
-        parent(Yₜ.c.turbconv[colidx]) .= FT(0)
-        parent(Yₜ.f.turbconv[colidx]) .= FT(0)
-    end
-
-    vertical_transport!(
-        Yₜ.c.ρ[colidx],
-        ᶠw[colidx],
-        ᶜρ[colidx],
-        ᶜρ[colidx],
-        dt,
-        Val(:none),
-    )
-
-    if :ρθ in propertynames(Y.c)
-        vertical_transport!(
-            Yₜ.c.ρθ[colidx],
-            ᶠw[colidx],
-            ᶜρ[colidx],
-            Y.c.ρθ[colidx],
-            dt,
-            energy_upwinding,
-        )
-    elseif :ρe_tot in propertynames(Y.c)
-        (; ᶜρh) = p
-        @. ᶜρh[colidx] = Y.c.ρe_tot[colidx] + ᶜp[colidx]
-        vertical_transport!(
-            Yₜ.c.ρe_tot[colidx],
-            ᶠw[colidx],
-            ᶜρ[colidx],
-            ᶜρh[colidx],
-            dt,
-            energy_upwinding,
-        )
-    elseif :ρe_int in propertynames(Y.c)
-        (; ᶜρh) = p
-        @. ᶜρh[colidx] = Y.c.ρe_int[colidx] + ᶜp[colidx]
-        vertical_transport!(
-            Yₜ.c.ρe_int[colidx],
-            ᶠw[colidx],
-            ᶜρ[colidx],
-            ᶜρh[colidx],
-            dt,
-            energy_upwinding,
-        )
-        @. Yₜ.c.ρe_int[colidx] += ᶜinterp(
-            dot(ᶠgradᵥ(ᶜp[colidx]), Geometry.Contravariant3Vector(ᶠw[colidx])),
-        )
-    end
-
-    Yₜ.c.uₕ[colidx] .= Ref(zero(eltype(Yₜ.c.uₕ[colidx])))
-
-    @. Yₜ.f.w[colidx] =
-        -(ᶠgradᵥ(ᶜp[colidx]) / ᶠinterp(ᶜρ[colidx]) + ᶠgradᵥ_ᶜΦ[colidx])
-    if p.tendency_knobs.rayleigh_sponge
-        @. Yₜ.f.w[colidx] -= p.ᶠβ_rayleigh_w[colidx] * Y.f.w[colidx]
-    end
-
-    for ᶜρc_name in filter(CA.is_tracer_var, propertynames(Y.c))
-        ᶜρcₜ = getproperty(Yₜ.c, ᶜρc_name)
-        ᶜρc = getproperty(Y.c, ᶜρc_name)
-        vertical_transport!(
-            ᶜρcₜ[colidx],
-            ᶠw[colidx],
-            ᶜρ[colidx],
-            ᶜρc[colidx],
-            dt,
-            tracer_upwinding,
-        )
-    end
-    return nothing
 end
 
 function remaining_tendency!(Yₜ, Y, p, t)
