+++ src/ClimaAtmos.jl
@@ -21,6 +21,8 @@ include(joinpath("tendencies", "rayleigh_sponge.jl"))
 include(joinpath("tendencies", "viscous_sponge.jl"))
 include(joinpath("tendencies", "advection.jl"))
 
+include(joinpath("tendencies", "implicit_tendency.jl"))
+
 include("model_getters.jl") # high-level (using parsed_args) model getters
 
 end # module
